udma: support receive inline acceleration.
category: feature
bugzilla: https://gitee.com/openeuler/umdk/issues/ID45DE
CVE: NA

-----------------------------------------------------------

When the inline acceleration function on the RX side is enabled,
the small packet data received by the chip is stored in the CQE
instead of the DDR space specified by the user. In the poll JFC
phase, the driver copies the data to the DDR space specified by
the user.

Signed-off-by: Jiahao Zhang <zhangjiahao14@huawei.com>