udma: support modify JFC.
category: feature
bugzilla: https://gitee.com/openeuler/umdk/issues/ID45DE
CVE: NA

-----------------------------------------------------------

The JFC attributes that can be modified include the number of
CQEs for the aggregation report and the period of the
aggregation report. The two parameters can be used together.
If either of the two conditions is met, the CEQE is reported.
Number of CQEs for the aggregation report: When the number of
CQEs is greater than the number of CQEs for the aggregation
report and arm_st is met, a CEQE event is reported.
period of the aggregation report: If the period since the
last CEQE event is longer than the period of the aggregation
report and arm_st is met, a CEQE event is reported.

Signed-off-by: Jiahao Zhang <zhangjiahao14@huawei.com>